Switched all if-else statements to patter matching

+++ src/main/scala/com/app/adv/models/Deduction.scala
@@ -1,3 +1,3 @@
 package com.app.adv.models
 
-case class Deduction(amount: BigDecimal)
+case class Deduction(amount: Double)
